Update patch set 3
Patch Set 3:

Do we need libwebm.lib to be useable before this can land? We don't release library files anywhere, so I didn't think it was a big deal in the near term.

I'll create a new libwebm project that's vs2013 native (instead of converted) soonish, and hopefully that'll link properly. Once that's done I can swing back around to webm_info and update the projects to use the library instead of sources directly. 

Anyway, not a big deal either way. Only bringing this up because I need to rebase the change that adds cue point support to webm_info if you don't think this is OK as a temporary solution.

Patch-set: 3