Update patch set 1
Patch Set 1:

> would this be called if the orientation were allowed to be changed or the view changed in some other way? or would it need to be relocated again?

If switching between portrait and landscape were allowed it would need to be.

No updates are required for landscape mode changes because the viewport's relative coordinates remain the same for both landscape modes because the square and texture vertices are the same for both landscape modes. That's free: Because GlkViewController, I think.

Patch-set: 1