Update patch set 1
Patch Set 1:

> so is there a common place where it should go to avoid the move? or how to respond to that is unknown right now?

Unknown right now, but I'm going to try enabling portrait mode orientations today and see what needs to be done. Ok with landing this, and then iterating on the portrait side? I actually think it's a reasonable expectation that this work in portrait mode (and I'm sure it would be handy for other openGL newbs, i.e. me, to not have to figure it out on their own)

Patch-set: 1