Update patch set 1
Patch Set 1:

Err... wait, I forgot: It's not as easy as just supporting portrait in openGL. I would also have to write code to manually reposition UI elements on the front page of the app. 

So, uh... not today, but maybe after I get a first usable version of this thing ready. OK with this as it is with that stipulation?

Patch-set: 1