alpha_encoder: Add support for VP9
alpha_encoder is broken as it is. It uses vpxenc for encoding
which defaults to vp9 and libwebm for muxing which uses vp8 as the
default video codec. As a result it creates a file with vp9 frames
but the header will say vp8.

This patch fixes all that mess and lets you create both vp8 and vp9
files with alpha by passing a -c parameter.

Change-Id: I6158deb39c1ad8574397973228e19e3486948ce0